refactor: decouple navigation and deep link logic into specialized helper files
This commit refactors the navigation logic within the `ui-logic` module by extracting specialized functionality from `DeepLinkHelper.kt` into separate files. This improves maintainability and follows the single responsibility principle.

Key changes:
- **`NavigationRouteHelper.kt`**: Created to house generic composable argument generation, deep link URI creation, and navigation link/intent generation logic.
- **`IntentAction.kt` & `IntentActionHelper.kt`**: Extracted `IntentAction` data structures and handling logic, specifically for managing `DC_API` (Digital Credential API) actions.
- **`DeepLinkAction.kt`**: Extracted `DeepLinkAction` and `DeepLinkType` definitions, including the parsing logic for various OpenID4VP, Credential Offer, and RQES schemes.
- **`DeepLinkHelper.kt`**: Refactored to focus exclusively on routing deep link actions to the appropriate application screens or broadcasting system intents.
- **Code Cleanup**: Applied consistent formatting and named arguments across the modified navigation helpers.
